Lots of cleanup, removals of 'this'.

+++ Roadie.Api.Library/Inspect/Plugins/Directory/DeleteUnwantedFiles.cs
@@ -2,7 +2,6 @@
 using Roadie.Library.Caching;
 using Roadie.Library.Configuration;
 using Roadie.Library.MetaData.ID3Tags;
-using Roadie.Library.Utility;
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -12,21 +11,16 @@ namespace Roadie.Library.Inspect.Plugins.Directory
 {
     public class DeleteUnwantedFiles : FolderPluginBase
     {
+        public override string Description => "Cleanup: Delete unwanted files";
+
         public override bool IsEnabled => true;
 
         public override bool IsPostProcessingPlugin => true;
 
-        public override string Description
-        {
-            get
-            {
-                return "Cleanup: Delete unwanted files";
-            }
-        }
-
         public override int Order { get; } = 99;
 
-        public DeleteUnwantedFiles(IRoadieSettings configuration, ICacheManager cacheManager, ILogger logger, IID3TagsHelper tagsHelper)
+        public DeleteUnwantedFiles(IRoadieSettings configuration, ICacheManager cacheManager, ILogger logger,
+                                            IID3TagsHelper tagsHelper)
             : base(configuration, cacheManager, logger, tagsHelper)
         {
         }
@@ -37,20 +31,17 @@ namespace Roadie.Library.Inspect.Plugins.Directory
             var data = string.Empty;
 
             var deletedFiles = new List<string>();
-            var fileExtensionsToDelete = this.Configuration.FileExtensionsToDelete ?? new string[0];
+            var fileExtensionsToDelete = Configuration.FileExtensionsToDelete ?? new string[0];
             foreach (var file in directory.GetFiles("*.*", SearchOption.AllDirectories))
-            {
                 if (fileExtensionsToDelete.Any(x => x.Equals(file.Extension, StringComparison.OrdinalIgnoreCase)))
                 {
-                    if (!Configuration.Inspector.IsInReadOnlyMode)
-                    {
-                        file.Delete();
-                    }
+                    if (!Configuration.Inspector.IsInReadOnlyMode) file.Delete();
                     deletedFiles.Add(file.Name);
                     Console.WriteLine($" X Deleted File [{file}], Was foud in in FileExtensionsToDelete");
                 }
-            }
-            result.Data = $"Deleted [{ deletedFiles.Count() }] unwanted files"; ;
+
+            result.Data = $"Deleted [{deletedFiles.Count()}] unwanted files";
+            ;
             result.IsSuccess = true;
             return result;
         }
