+++ Roadie.Api.Library/Processors/FileProcessor.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.Extensions.Logging;
+using MimeMapping;
 using Roadie.Library.Caching;
 using Roadie.Library.Configuration;
 using Roadie.Library.Data;
@@ -8,7 +9,6 @@ using Roadie.Library.Extensions;
 using Roadie.Library.Factories;
 using Roadie.Library.FilePlugins;
 using Roadie.Library.MetaData.Audio;
-using Roadie.Library.MetaData.ID3Tags;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
@@ -20,13 +20,13 @@ namespace Roadie.Library.Processors
 {
     public sealed class FileProcessor : ProcessorBase
     {
-        private IEnumerable<IFilePlugin> _plugins = null;
+        private IEnumerable<IFilePlugin> _plugins;
 
         public IEnumerable<IFilePlugin> Plugins
         {
             get
             {
-                if (this._plugins == null)
+                if (_plugins == null)
                 {
                     var plugins = new List<IFilePlugin>();
                     try
@@ -35,55 +35,53 @@ namespace Roadie.Library.Processors
                         var types = AppDomain.CurrentDomain.GetAssemblies()
                             .SelectMany(s => s.GetTypes())
                             .Where(p => type.IsAssignableFrom(p));
-                        foreach (Type t in types)
-                        {
+                        foreach (var t in types)
                             if (t.GetInterface("IFilePlugin") != null && !t.IsAbstract && !t.IsInterface)
                             {
-      
-
-                                IFilePlugin plugin = Activator.CreateInstance(t, new object[] { this.Configuration, this.HttpEncoder, this.ArtistFactory, this.ReleaseFactory,
-                                                                                                this.ImageFactory, this.CacheManager, this.Logger, this.ArtistLookupEngine,
-                                                                                                this.ReleaseLookupEngine, this.AudioMetaDataHelper}) as IFilePlugin;
+                                var plugin = Activator.CreateInstance(t, Configuration, HttpEncoder, ArtistFactory,
+                                    ReleaseFactory, ImageFactory, CacheManager, Logger, ArtistLookupEngine,
+                                    ReleaseLookupEngine, AudioMetaDataHelper) as IFilePlugin;
                                 plugins.Add(plugin);
                             }
-                        }
                     }
                     catch (Exception ex)
                     {
-                        this.Logger.LogError(ex);
+                        Logger.LogError(ex);
                     }
-                    this._plugins = plugins.ToArray();
+
+                    _plugins = plugins.ToArray();
                 }
-                return this._plugins;
+
+                return _plugins;
             }
         }
 
-        public FileProcessor(IRoadieSettings configuration, IHttpEncoder httpEncoder, string destinationRoot, IRoadieDbContext context, ICacheManager cacheManager, ILogger logger, IArtistLookupEngine artistLookupEngine, IArtistFactory artistFactory, IReleaseFactory releaseFactory, IImageFactory imageFactory, IReleaseLookupEngine releaseLookupEngine, IAudioMetaDataHelper audioMetaDataHelper )
-            : base(configuration, httpEncoder, destinationRoot, context, cacheManager, logger, artistLookupEngine, artistFactory, releaseFactory, imageFactory, releaseLookupEngine, audioMetaDataHelper)
+        public FileProcessor(IRoadieSettings configuration, IHttpEncoder httpEncoder, string destinationRoot,
+                    IRoadieDbContext context, ICacheManager cacheManager, ILogger logger,
+            IArtistLookupEngine artistLookupEngine, IArtistFactory artistFactory, IReleaseFactory releaseFactory,
+            IImageFactory imageFactory, IReleaseLookupEngine releaseLookupEngine,
+            IAudioMetaDataHelper audioMetaDataHelper)
+            : base(configuration, httpEncoder, destinationRoot, context, cacheManager, logger, artistLookupEngine,
+                artistFactory, releaseFactory, imageFactory, releaseLookupEngine, audioMetaDataHelper)
         {
         }
 
-        public static string DetermineFileType(System.IO.FileInfo fileinfo)
+        public static string DetermineFileType(FileInfo fileinfo)
         {
-            string r = MimeMapping.MimeUtility.GetMimeMapping(fileinfo.FullName);
+            var r = MimeUtility.GetMimeMapping(fileinfo.FullName);
             if (r.Equals("application/octet-stream"))
             {
-                if (fileinfo.Extension.Equals(".cue"))
-                {
-                    r = "audio/r-cue";
-                }
-                if (fileinfo.Extension.Equals(".mp4") || fileinfo.Extension.Equals(".m4a"))
-                {
-                    r = "audio/mp4";
-                }
+                if (fileinfo.Extension.Equals(".cue")) r = "audio/r-cue";
+                if (fileinfo.Extension.Equals(".mp4") || fileinfo.Extension.Equals(".m4a")) r = "audio/mp4";
             }
+
             Trace.WriteLine(string.Format("FileType [{0}] For File [{1}]", r, fileinfo.FullName));
             return r;
         }
 
         public async Task<OperationResult<bool>> Process(string filename, bool doJustInfo = false)
         {
-            return await this.Process(new FileInfo(filename), doJustInfo);
+            return await Process(new FileInfo(filename), doJustInfo);
         }
 
         public async Task<OperationResult<bool>> Process(FileInfo fileInfo, bool doJustInfo = false)
@@ -93,76 +91,70 @@ namespace Roadie.Library.Processors
             try
             {
                 // Determine what type of file this is
-                var fileType = FileProcessor.DetermineFileType(fileInfo);
+                var fileType = DetermineFileType(fileInfo);
 
                 OperationResult<bool> pluginResult = null;
-                foreach (var p in this.Plugins)
-                {
+                foreach (var p in Plugins)
                     // See if there is a plugin
                     if (p.HandlesTypes.Contains(fileType))
                     {
-                        pluginResult = await p.Process(this.DestinationRoot, fileInfo, doJustInfo, this.SubmissionId);
+                        pluginResult = await p.Process(DestinationRoot, fileInfo, doJustInfo, SubmissionId);
                         break;
                     }
-                }
 
                 if (!doJustInfo)
-                {
                     // If no plugin, or if plugin not successfull and toggle then move unknown file
-                    if ((pluginResult == null || !pluginResult.IsSuccess) && this.DoMoveUnknowns)
+                    if ((pluginResult == null || !pluginResult.IsSuccess) && DoMoveUnknowns)
                     {
-                        var uf = this.UnknownFolder;
+                        var uf = UnknownFolder;
                         if (!string.IsNullOrEmpty(uf))
                         {
-                            if (!Directory.Exists(uf))
-                            {
-                                Directory.CreateDirectory(uf);
-                            }
-                            if (!fileInfo.DirectoryName.Equals(this.UnknownFolder))
-                            {
+                            if (!Directory.Exists(uf)) Directory.CreateDirectory(uf);
+                            if (!fileInfo.DirectoryName.Equals(UnknownFolder))
                                 if (File.Exists(fileInfo.FullName))
                                 {
-                                    var df = Path.Combine(this.UnknownFolder, string.Format("{0}~{1}~{2}", Guid.NewGuid(), fileInfo.Directory.Name, fileInfo.Name));
-                                    this.Logger.LogDebug("Moving Unknown/Invalid File [{0}] -> [{1}] to UnknownFolder", fileInfo.FullName, df);
+                                    var df = Path.Combine(UnknownFolder,
+                                        string.Format("{0}~{1}~{2}", Guid.NewGuid(), fileInfo.Directory.Name,
+                                            fileInfo.Name));
+                                    Logger.LogDebug("Moving Unknown/Invalid File [{0}] -> [{1}] to UnknownFolder",
+                                        fileInfo.FullName, df);
                                     fileInfo.MoveTo(df);
                                 }
-                            }
                         }
                     }
-                }
+
                 result = pluginResult;
             }
-            catch (System.IO.PathTooLongException ex)
+            catch (PathTooLongException ex)
             {
-                this.Logger.LogError(ex, string.Format("Error Processing File. File Name Too Long. Deleting."));
-                if (!doJustInfo)
-                {
-                    fileInfo.Delete();
-                }
+                Logger.LogError(ex, "Error Processing File. File Name Too Long. Deleting.");
+                if (!doJustInfo) fileInfo.Delete();
             }
             catch (Exception ex)
             {
-                var willMove = !fileInfo.DirectoryName.Equals(this.UnknownFolder);
-                this.Logger.LogError(ex, string.Format("Error Processing File [{0}], WillMove [{1}]\n{2}", fileInfo.FullName, willMove, ex.Serialize()));
+                var willMove = !fileInfo.DirectoryName.Equals(UnknownFolder);
+                Logger.LogError(ex,
+                    string.Format("Error Processing File [{0}], WillMove [{1}]\n{2}", fileInfo.FullName, willMove,
+                        ex.Serialize()));
                 string newPath = null;
                 try
                 {
-                    newPath = Path.Combine(this.UnknownFolder, fileInfo.Directory.Parent.Name, fileInfo.Directory.Name, fileInfo.Name);
+                    newPath = Path.Combine(UnknownFolder, fileInfo.Directory.Parent.Name, fileInfo.Directory.Name,
+                        fileInfo.Name);
                     if (willMove && !doJustInfo)
                     {
                         var directoryPath = Path.GetDirectoryName(newPath);
-                        if (!Directory.Exists(directoryPath))
-                        {
-                            Directory.CreateDirectory(directoryPath);
-                        }
+                        if (!Directory.Exists(directoryPath)) Directory.CreateDirectory(directoryPath);
                         fileInfo.MoveTo(newPath);
                     }
                 }
                 catch (Exception ex1)
                 {
-                    this.Logger.LogError(ex1, string.Format("Unable to move file [{0}] to [{1}]", fileInfo.FullName, newPath));
+                    Logger.LogError(ex1,
+                        string.Format("Unable to move file [{0}] to [{1}]", fileInfo.FullName, newPath));
                 }
             }
+
             return result;
         }
     }
